# src/main/resources/openapi/lscrp_oas_0.2.5.yaml
openapi: 3.0.0
info:
  title: County RP API
  description: API
  version: 1.0.0
paths:
  /account:
    post:
      operationId: createAccount
      tags:
        - accountApi
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    get:
      operationId: listAccount
      tags:
        - accountApi
      summary: Get a list of all accounts
      security:
        - bearerAuth: [ ]
      parameters:
        - in: query
          name: username
          schema:
            type: string
          required: false
          description: Filter accounts by username
        - in: query
          name: emailAddress
          schema:
            type: string
          required: false
          description: Filter accounts by email address
      x-roles-allowed:
        - user
        - admin
        - super_admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '500':
          description: Internal Server Error
  /account/{accountId}:
    get:
      operationId: retrieveAccount
      tags:
        - accountApi
      summary: Get an account by ID
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

    patch:
      operationId: patchAccount
      tags:
        - accountApi
      summary: Update an account by ID
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

    delete:
      operationId: deleteAccount
      tags:
        - accountApi
      summary: Delete an account by ID
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /account/{accountId}/logs/login:
    get:
      operationId: listAccountLoginLogs
      tags:
        - accountApi
      summary: Get a list of all account login logs
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int32
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoginLogAudit'
        '500':
          description: Internal Server Error
  /character:
    post:
      operationId: createCharacter
      tags:
        - characterApi
      summary: Create a new character
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
    get:
      summary: Get a list of characters
      operationId: listCharacters
      tags:
        - characterApi
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: Filter character by name
        - in: query
          name: accountUsername
          schema:
            type: string
          required: false
          description: Filter characters by account username
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
  /character/unapproved:
    get:
      summary: Get a list of unapproved characters
      operationId: listUnapprovedCharacter
      tags:
        - characterApi
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - admin
        - super_admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
  /character/{id}:
    get:
      summary: Get a character by ID
      operationId: retrieveCharacter
      tags:
        - characterApi
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
    patch:
      summary: Update a character by ID
      operationId: patchCharacter
      tags:
        - characterApi
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - admin
        - super_admin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
    delete:
      summary: Delete a character by ID
      operationId: deleteCharacter
      tags:
        - characterApi
      security:
        - bearerAuth: [ ]
      x-roles-allowed:
        - user
        - super_admin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
        email:
          type: string
        isInGame:
          type: boolean
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        lastLoginAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        admin:
          type: integer
        referralCode:
          type: string
        premium:
          type: integer
        premiumCoins:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - isInGame
        - email
        - admin
        - characters
        - premium
        - premiumCoins
        - createdAt
    AccountCreate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: string
      required:
        - username
        - email
        - password
    AccountUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: string
    Character:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        birthday:
          type: string
          format: date
        accountId:
          type: integer
          format: int32
        accountUsername:
          type: string
        gender:
          $ref: '#/components/schemas/CharacterGender'
        health:
          type: integer
          minimum: 0
          maximum: 100
        armour:
          type: integer
          minimum: 0
          maximum: 100
        skin:
          type: integer
        money:
          type: integer
        paycheck:
          type: integer
        bank:
          type: integer
        savings:
          type: integer
        hours:
          type: integer
        minutes:
          type: integer
        level:
          type: integer
        positionX:
          type: number
        positionY:
          type: number
        positionZ:
          type: number
        angle:
          type: number
        virtualWorld:
          type: integer
        interior:
          type: integer
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        factionId:
          type: integer
        rankId:
          type: integer
        isLeader:
          type: boolean
        job:
          type: integer
        state:
          type: integer
        maskId:
          type: string
        respawnTime:
          type: integer
        drugAddiction:
          type: integer
        walkingStyle:
          type: integer
          description: The walking style of the character
        fightingStyle:
          type: integer
          description: The fighting style of the character
        isApproved:
          type: boolean
        approvedBy:
          type: string
        deniedReason:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - accountId
        - accountUsername
        - name
        - birthday
        - gender
        - health
        - armour
        - skin
        - money
        - paycheck
        - bank
        - savings
        - hours
        - minutes
        - level
        - positionX
        - positionY
        - positionZ
        - angle
        - virtualWorld
        - interior
        - isLeader
        - job
        - state
        - respawnTime
        - drugAddiction
        - walkingStyle
        - fightingStyle
        - createdAt
        - factionId
    CharacterCreate:
      type: object
      properties:
        accountId:
          type: integer
          format: int32
        name:
          type: string
        birthday:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/CharacterGender'
        skin:
          type: integer
      required:
        - accountId
        - name
        - birthday
        - gender
    CharacterGender:
      type: string
      enum: [ MALE, FEMALE ]
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
        health:
          type: integer
          minimum: 0
          maximum: 100
        skin:
          type: integer
        armour:
          type: integer
          minimum: 0
          maximum: 100
        isApproved:
          type: boolean
        approvedBy:
          type: string
        deniedReason:
          type: string
        approvedAt:
          type: string
          format: date-time
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        model:
          type: integer
          description: Model ID of the vehicle
        owner:
          $ref: '#/components/schemas/Character'
        price:
          type: integer
          format: int32
        health:
          type: number
        numberplate:
          type: string
        positionX:
          type: number
          format: float
        positionY:
          type: number
          format: float
        positionZ:
          type: number
          format: float
        rotation:
          type: number
          format: float
        parkPositionX:
          type: number
          format: float
        parkPositionY:
          type: number
          format: float
        parkPositionZ:
          type: number
          format: float
        parkRotation:
          type: number
          format: float
        primaryColor:
          type: integer
          description: Primary color code
        secondaryColor:
          type: integer
          description: Secondary color code
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - model
        - price
        - owner
        - positionX
        - positionY
        - positionZ
        - rotation
        - primaryColor
        - secondaryColor
        - parkPositionX
        - parkPositionY
        - parkPositionZ
        - parkRotation
...
    Property:
      type: object
      properties:
        id:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/PropertyType'
        bizType:
          $ref: '#/components/schemas/BizType'
        name:
          type: string
          maxLength: 128
        owner:
          type: string
        price:
          type: integer
        address:
          type: string
          maxLength: 64
        positionX:
          type: number
          format: float
        positionY:
          type: number
          format: float
        positionZ:
          type: number
          format: float
        exteriorVirtualWorld:
          type: integer
        exterior:
          type: integer
        interiorPositionX:
          type: number
          format: float
        interiorPositionY:
          type: number
          format: float
        interiorPositionZ:
          type: number
          format: float
        interior:
          type: integer
        isLocked:
          type: boolean

      required:
        - name
        - price
        - positionX
        - positionY
        - positionZ
        - exteriorVirtualWorld
        - exterior
        - interiorPositionX
        - interiorPositionY
        - interiorPositionZ
        - interior
        - isLocked
...
    PropertyType:
      type: string
      enum: [ HOUSE, BUSINESS, APARTMENT, WAREHOUSE, GARAGE ]
    BizType:
      type: integer
      enum: [ 0, 1, 2, 3, 4, 5 ]
      x-enum-varnames: [ GROCERY_STORE, GAS_STATION, CLOTHING_STORE, OTHER, BAR, RESTAURANT ]
...
    LoginLogAudit:
      type: object
      properties:
        accountId:
          type: integer
          format: int32
        ipAddress:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - ipAddress
        - createdAt
        - success
...
security:
  - bearerAuth: [ ]